fix(extensions): moved img parts to yfm preset
