--- src/extensions/behavior/ImgSizeAdditions/const.ts
@@ -1,5 +1,0 @@
-import {PluginKey} from 'prosemirror-state';
-
-export const IMG_MAX_HEIGHT = 600; //px
-export type ImageRendererState = {linkAdded: boolean};
-export const imageRendererKey = new PluginKey<ImageRendererState>('imageRenderer');
