+++ src/extensions/yfm/ImgSize/ImagePaste/skeleton.scss
