+++ src/extensions/yfm/ImgSize/ImageWidget/view.scss
